Refactor: add artworkHelper and move related function to there
